Rewrite view to be more coherent
Part of this rewrite was renaming box.html to
_box.html to be in line with the convention of
naming partials.

In addition, the ajax portion will return a new
partial called _box_body.html. The templates that
ship with this are effectively the same as the
original box.html, however this separation gives
more flexibility in how ajax boxes can work.